Added in AnimationLibrary
