+++ SquadIO/SquadIO.vcxproj
@@ -86,7 +86,7 @@
     </Link>
   </ItemDefinitionGroup>
   <ItemGroup>
-    <ClInclude Include="AnimationManager.h" />
+    <ClInclude Include="include\AnimationLibrary.h" />
     <ClInclude Include="include\AudioManager.h" />
     <ClInclude Include="include\AudioResource.h" />
     <ClInclude Include="include\EngineObject.h" />
@@ -118,6 +118,7 @@
     <ClInclude Include="zlib\include\zlib.h" />
   </ItemGroup>
   <ItemGroup>
+    <ClCompile Include="src\AnimationLibrary.cpp" />
     <ClCompile Include="src\AudioManager.cpp" />
     <ClCompile Include="src\AudioResource.cpp" />
     <ClCompile Include="src\EngineObject.cpp" />
