+++ SquadIO/SquadIO.vcxproj.filters
@@ -76,6 +76,9 @@
     <ClCompile Include="src\network\DeliveryNotificationManager.cpp">
       <Filter>Source Files\network</Filter>
     </ClCompile>
+    <ClCompile Include="src\AnimationLibrary.cpp">
+      <Filter>Source Files\network</Filter>
+    </ClCompile>
   </ItemGroup>
   <ItemGroup>
     <ClInclude Include="include\network\NetworkManager.h">
@@ -164,6 +167,9 @@
     </ClInclude>
     <ClInclude Include="include\gameResource.h">
       <Filter>Header Files\Resource Classes</Filter>
+    </ClInclude>
+    <ClInclude Include="include\AnimationLibrary.h">
+      <Filter>Header Files\Render Classes</Filter>
     </ClInclude>
   </ItemGroup>
   <ItemGroup>
